Moving gamepad USB hosts to their own folder
