+++ headers/hosts/Ultrastik360Host.h
@@ -1,0 +1,38 @@
+/*
+ * SPDX-License-Identifier: MIT
+ * SPDX-FileCopyrightText: Copyright (c) 2026 OpenStickCommunity (gp2040-ce.info)
+ */
+
+#ifndef _ULTRASTIK360HOST_H_
+#define _ULTRASTIK360PROHOST_H_
+
+#include <gphost.h>
+
+// Ultrastik 360
+typedef struct TU_ATTR_PACKED
+{
+    uint8_t GD_GamePadPointerX;
+    uint8_t GD_GamePadPointerY;
+
+    struct {
+        uint8_t BTN_GamePadButton1 : 1;
+        uint8_t BTN_GamePadButton2 : 1;
+        uint8_t BTN_GamePadButton3 : 1;
+        uint8_t BTN_GamePadButton4 : 1;
+        uint8_t BTN_GamePadButton5 : 1;
+        uint8_t BTN_GamePadButton6 : 1;
+        uint8_t BTN_GamePadButton7 : 1;
+        uint8_t BTN_GamePadButton8 : 1;
+        uint8_t BTN_GamePadButton9 : 1;
+        uint8_t BTN_GamePadButton10 : 1;
+        uint8_t BTN_GamePadButton11 : 1;
+        uint8_t BTN_GamePadButton12 : 1;
+        uint8_t BTN_GamePadButton13 : 1;
+        uint8_t BTN_GamePadButton14 : 1;
+        uint8_t BTN_GamePadButton15 : 1;
+        uint8_t padding : 1;
+    };
+
+} ultrastik360_t;
+
+#endif
